% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet2GatingML.R
\name{addCustomInfo}
\alias{addCustomInfo}
\title{add customInfo nodes to each gate node and add BooleanAndGates}
\usage{
addCustomInfo(root, gs, flowEnv, cytobank.default.scale = TRUE, showHidden)
}
\description{
add customInfo nodes to each gate node and add BooleanAndGates
}

