% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\name{ce_get_compensations}
\alias{ce_get_compensations}
\title{Obtain the spillover matrices for the samples in a Cytobank experiment}
\usage{
ce_get_compensations(x)
}
\arguments{
\item{x}{A cytobank_experiment object}
}
\value{
A named list of spillover matrices
}
\description{
Obtain the spillover matrices for the samples in a Cytobank experiment
}
