% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\name{ce_get_markers}
\alias{ce_get_markers}
\title{Extract markers from cytobank_experiment}
\usage{
ce_get_markers(x, panel_name = NULL)
}
\arguments{
\item{x}{A \code{cytobank_experiment} object}

\item{panel_name}{select panel to process}
}
\description{
Extract markers from cytobank_experiment
}
