% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\name{ce_get_panels}
\alias{ce_get_panels}
\title{Obtain counts of the number of samples associated with each marker panel
in a Cytobank experiment}
\usage{
ce_get_panels(x)
}
\arguments{
\item{x}{\code{cytobank_experiment} object}
}
\value{
A \code{tibble} of panels with sample counts
}
\description{
Obtain counts of the number of samples associated with each marker panel
in a Cytobank experiment
}
