% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\name{ce_get_samples}
\alias{ce_get_samples}
\title{Obtain a mapping between the samples and marker panels in a Cytobank
experiment}
\usage{
ce_get_samples(x)
}
\arguments{
\item{x}{A \code{cytobank_experiment} object}
}
\value{
A \code{tibble} with rows containing sample names and their associated
panel names
}
\description{
Obtain a mapping between the samples and marker panels in a Cytobank
experiment
}
