% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMerge.R
\name{clone.GatingSet}
\alias{clone.GatingSet}
\title{light weight version of clone method
eventually will merge to flowWorkspace}
\usage{
clone.GatingSet(x, ...)
}
\description{
light weight version of clone method
eventually will merge to flowWorkspace
}

