% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGML_methods.R
\name{compensate,GatingSet,graphGML-method}
\alias{compensate,GatingSet,graphGML-method}
\title{compensate a GatingSet based on the compensation information stored in graphGML object}
\usage{
\S4method{compensate}{GatingSet,graphGML}(x, spillover, ...)
}
\arguments{
\item{x}{GatingSet}

\item{spillover}{graphGML}

\item{...}{unused.}
}
\value{
compensated GatingSet
}
\description{
compensate a GatingSet based on the compensation information stored in graphGML object
}
