% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\docType{methods}
\name{cytobank_experiment-methods}
\alias{cytobank_experiment-methods}
\alias{print.cytobank_experiment}
\alias{markernames,cytobank_experiment-method}
\alias{colnames,cytobank_experiment-method}
\alias{sampleNames,cytobank_experiment-method}
\alias{pData,cytobank_experiment-method}
\title{Methods for interacting with cytobank_experiment objects}
\usage{
markernames(object)

colnames(object)

sampleNames(object)

pData(object)
}
\arguments{
\item{object}{A \code{cytobank_experiment} object}
}
\description{
These methods mirror similar accessor methods for the \code{GatingSet}
class.
}
