% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\docType{methods}
\name{cytobankExperiment}
\alias{cytobankExperiment}
\alias{print.cytobankExperiment}
\alias{getCompensationMatrices.cytobankExperiment}
\alias{markernames,cytobankExperiment-method}
\alias{colnames,cytobankExperiment-method}
\alias{getTransformations.cytobankExperiment}
\alias{sampleNames,cytobankExperiment-method}
\alias{pData,cytobankExperiment-method}
\title{Construct cytobankExperiment object from ACS file}
\usage{
cytobankExperiment(acs, exdir = tempfile())

\method{print}{cytobankExperiment}(x)

\method{getCompensationMatrices}{cytobankExperiment}(x)

\S4method{markernames}{cytobankExperiment}(object)

\S4method{colnames}{cytobankExperiment}(x, do.NULL = "missing",
  prefix = "missing")

\method{getTransformations}{cytobankExperiment}(x)

\S4method{sampleNames}{cytobankExperiment}(object)

\S4method{pData}{cytobankExperiment}(object)
}
\arguments{
\item{acs}{ACS file exported from Cytobank}

\item{exdir}{he directory to extract files to}

\item{x}{cytobankExperiment object}

\item{object}{cytobankExperiment object}
}
\value{
cytobankExperiment object
}
\description{
Construct cytobankExperiment object from ACS file
}
