% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet2cytobank.R, R/GatingSet2flowJo.R,
%   R/cytobankExperiment.R, R/flowJoWorkspace_Methods.R, R/diva2GatingSet.R
\name{CytoML-deprecated}
\alias{CytoML-deprecated}
\title{Deprecated functions in package \pkg{CytoML}.}
\description{
\code{GatingSet2cytobank} --> \code{\link{gatingset_to_cytobank}}

\code{GatingSet2flowJo} --> \code{\link{gatingset_to_flowjo}}

\code{cytobankExperiment} --> \code{\link{open_cytobank_experiment}}

\code{cytobank2GatingSet} --> \code{\link{cytobank_to_gatingset}}

\code{parseWorkspace} --> \code{\link{flowjo_to_gatingset}}

\code{getKeywords} --> \code{\link{fj_ws_get_keywords}}

\code{getSamples} --> \code{\link{fj_ws_get_samples}}

\code{getSampleGroups} --> \code{\link{fj_ws_get_sample_groups}}

\code{openDiva} --> \code{\link{open_diva_xml}}

\code{parseWorkspace} --> \code{\link{diva_to_gatingset}}
}
