% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{CytoML.par.set}
\alias{CytoML.par.set}
\alias{CytoML.par.get}
\title{CytoML.par.set sets a set of parameters in the CytoML package namespace.}
\usage{
CytoML.par.set(name, value)

CytoML.par.get(name = NULL)
}
\arguments{
\item{name}{The name of a parameter category to get or set.}

\item{value}{A named list of values to set for category name or a list of such lists if name is missing.}
}
\description{
CytoML.par.get gets a set of parameters in the CytoML package namespace.
}
\details{
It is currently used to add/remove the support for a specific flowJo versions (parsed from xml node '/Workspace/version' in flowJo workspace)
}
\examples{
 #get the flowJo versions currently supported 
 old <- CytoML.par.get("flowJo_versions")

 #add the new version
 old[["win"]] <- c(old[["win"]], "1.7")    
 CytoML.par.set("flowJo_versions", old)
 
 CytoML.par.get("flowJo_versions")

}
