% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{CytoML.par.init}
\alias{CytoML.par.init}
\title{workspace version is parsed from xml node '/Workspace/version' in flowJo workspace
and matched with this list to dispatch to the one of the three workspace parsers}
\usage{
CytoML.par.init()
}
\description{
workspace version is parsed from xml node '/Workspace/version' in flowJo workspace
and matched with this list to dispatch to the one of the three workspace parsers
}
