% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva2GatingSet.R
\name{diva_get_samples}
\alias{diva_get_samples}
\alias{diva_get_sample_groups}
\title{Get a table of samples from a FACSDiva workspace}
\usage{
diva_get_samples(x)
diva_get_sample_groups(x)
}
\arguments{
\item{x}{A \code{diva_workspace}}
}
\value{
A \code{data.frame} with columns \code{tub}, \code{name}, and \code{specimen}
}
\description{
Return a data.frame of sample group information from a FACSDiva workspace
}
