% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva2GatingSet.R
\docType{class}
\name{divaWorkspace-class}
\alias{divaWorkspace-class}
\alias{getSamples,divaWorkspace-method}
\alias{getSampleGroups,divaWorkspace-method}
\alias{show,divaWorkspace-method}
\alias{parseWorkspace.divaWorkspace}
\title{divaWorkspace class
Inherited from \link{flowJoWorkspace-class}}
\usage{
\S4method{getSamples}{divaWorkspace}(x)

\S4method{getSampleGroups}{divaWorkspace}(x)

\S4method{show}{divaWorkspace}(object)

\method{parseWorkspace}{divaWorkspace}(obj, ...)
}
\arguments{
\item{x}{divaWorkspace}

\item{object}{divaWorkspace}

\item{obj}{divaWorkspace}

\item{...}{other arguments}
}
\description{
divaWorkspace class
Inherited from \link{flowJoWorkspace-class}
}
