% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowJoWorkspace_Methods.R
\docType{class}
\name{flowjo_workspace-class}
\alias{flowjo_workspace-class}
\alias{show,flowjo_workspace-method}
\title{An R representation of a flowJo workspace.}
\description{
Objects can be created by calls of the form \code{new("flowjo_workspace.xml", ...)}.
}
\section{Slots}{
 
\describe{
  \item{\code{doc}:}{Object of class \code{"externalptr"}. }
  }
}

\examples{
  require(flowWorkspaceData)
  d<-system.file("extdata",package="flowWorkspaceData")
  wsfile<-list.files(d,pattern="A2004Analysis.xml",full=TRUE)
  ws <- open_flowjo_xml(wsfile);
  ws
  fj_ws_get_samples(ws)

}
\seealso{
\code{\linkS4class{GatingSet}} 
  \code{\linkS4class{GatingHierarchy}}
}
