% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGML_methods.R
\name{gating_graphGML}
\alias{gating_graphGML}
\title{Apply the gatingML graph to a GatingSet}
\usage{
gating_graphGML(x, y, ...)
}
\arguments{
\item{x}{graphGML}

\item{y}{GatingSet}

\item{...}{other arguments}
}
\value{
Nothing. As the side effect, gates generated by gating methods are saved in \code{GatingSet}.
}
\description{
It applies the gates to the GatingSet based on the population tree described in graphGML.
}
