% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowJoWorkspace_Methods.R
\docType{methods}
\name{GatingSet,character,character-method}
\alias{GatingSet,character,character-method}
\alias{GatingSet}
\title{constructors for GatingSet}
\usage{
\S4method{GatingSet}{character,character}(x, y, guids,
  includeGates = FALSE, sampNloc = "keyword", xmlParserOption, wsType)
}
\arguments{
\item{x}{\code{character} or \code{flowSet} or \code{GatingHierarchy}}

\item{y}{\code{character} or\code{missing}}

\item{guids}{\code{character} vectors to uniquely identify each sample (Sometime FCS file names alone may not be unique)}

\item{includeGates}{\code{logical} whether to parse the gates or just simply extract the flowJo stats}

\item{sampNloc}{\code{character} scalar indicating where to get sampleName(or FCS filename) within xml workspace. It is either from "keyword" or "sampleNode".}

\item{xmlParserOption}{\code{integer} option passed to \code{\link{xmlTreeParse}}}

\item{wsType}{\code{character} workspace type, can be value of "win", "macII", "vX", "macIII".}
}
\description{
construct object from xml workspace file and a list of sampleIDs (not intended to be called by user.)
}
