% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet2GatingML.R
\name{GatingSet2Environment}
\alias{GatingSet2Environment}
\title{extract the trans, comps and gates into GML2 compatible format
and save to environment}
\usage{
GatingSet2Environment(gs, cytobank.default.scale = TRUE, showHidden)
}
\description{
extract the trans, comps and gates into GML2 compatible format
and save to environment
}

