% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet2GatingML.R
\name{GatingSet2GatingML}
\alias{GatingSet2GatingML}
\title{Convert a GatingSet to gatingML}
\usage{
GatingSet2GatingML(gs, output, showHidden = FALSE, type = c("cytobank",
  "flowjo"), ...)
}
\arguments{
\item{output}{a file name when type is "cytobank", a folder name when type is "flowjo"}

\item{showHidden}{whether to include the hidden population nodes in the output}

\item{type}{either "cytobank" or "flowjo"}

\item{...}{other arguments}

\item{a}{GatingSet object}
}
\description{
this function retrieves the gates from GatingSet
and writes a customed GatingML-2.0 file
which can then be imported into cytobank
operation:
1. Read in gate geometry, compensation and transformation from gatingSet
2. Rescale gate boundaries with flowJoTrans() so gates show up in flowJo
3. Save gates and hierarchy structure to R environment
4. Write environment out to gatingML using write.GatingML()
}
\examples{
\dontrun{
library(CytoML)
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
outFile <- tempfile(fileext = ".xml")
GatingSet2GatingML(gs, outFile)
}
}

