% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGML-methods.R
\docType{methods}
\name{getChildren,graphGML,character-method}
\alias{getChildren,graphGML,character-method}
\title{get children nodes}
\usage{
\S4method{getChildren}{graphGML,character}(obj, y)
}
\arguments{
\item{obj}{\code{graphGML}}

\item{y}{\code{character} parent node path}
}
\description{
get children nodes
}
\examples{
\dontrun{
xmlfile <- system.file("extdata/cytotrol_tcell_cytobank.xml", package = "CytoML")
g <- read.gatingML.cytobank(xmlfile)
getChildren(g, "GateSet_722326")
getParent(g, "GateSet_722326")
}
}

