% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGML_methods.R
\name{getCompensationMatrices.graphGML}
\alias{getCompensationMatrices.graphGML}
\title{Extract compensation from graphGML object.}
\usage{
\method{getCompensationMatrices}{graphGML}(x)
}
\arguments{
\item{x}{graphGML}
}
\value{
compensation object or "FCS" when compensation comes from FCS keywords
}
\description{
Extract compensation from graphGML object.
}
