% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGML_methods.R
\name{getGate,graphGML,character-method}
\alias{getGate,graphGML,character-method}
\title{get gate from the node}
\usage{
\S4method{getGate}{graphGML,character}(obj, y)
}
\arguments{
\item{obj}{\code{graphGML}}

\item{y}{\code{character} node path}
}
\value{
the gate information associated with the node
}
\description{
get gate from the node
}
