% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGML_methods.R
\name{getParent,graphGML,character-method}
\alias{getParent,graphGML,character-method}
\title{get parent nodes}
\usage{
\S4method{getParent}{graphGML,character}(obj, y)
}
\arguments{
\item{obj}{\code{graphGML}}

\item{y}{\code{character} child node path}
}
\value{
a graphNEL node
}
\description{
get parent nodes
}
