% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowJoWorkspace_Methods.R
\name{getSamples}
\alias{getSamples}
\alias{fj_ws_get_samples}
\title{Get a list of samples from a flowJo workspace}
\usage{
getSamples(...)

fj_ws_get_samples(x, sampNloc = "keyword")
}
\arguments{
\item{x}{A \code{flowjo_workspace}}

\item{sampNloc}{\code{character} either "keyword" or "sampleNode". see \link{flowjo_to_gatingset}}
}
\value{
A \code{data.frame} with columns \code{sampleID}, \code{name}, and \code{compID} if \code{x} is a \code{flowjo_workspace}.
}
\description{
Return  a data frame of samples contained in a flowJo workspace
}
\details{
Returns a \code{data.frame} of samples in the \code{flowjo_workspace}, including their 
\code{sampleID}, \code{name}, and \code{compID} (compensation matrix ID).
}
\examples{
      \dontrun{
        #ws is a flowjo_workspace
        getSamples(ws);
      }
}
