% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGML_methods.R
\name{getTransformations.graphGML}
\alias{getTransformations.graphGML}
\title{Extract transformations from graphGML object.}
\usage{
\method{getTransformations}{graphGML}(x, ...)
}
\arguments{
\item{x}{graphGML}

\item{...}{not used}
}
\value{
transformerList object
}
\description{
Extract transformations from graphGML object.
}
