% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\name{cytobankExperiment}
\alias{cytobankExperiment}
\alias{open_cytobank_experiment}
\title{Construct cytobank_experiment object from ACS file}
\usage{
cytobankExperiment(...)

open_cytobank_experiment(acs, exdir = tempfile())
}
\arguments{
\item{acs}{ACS file exported from Cytobank}

\item{exdir}{he directory to extract files to}
}
\value{
cytobank_experiment object
}
\description{
Construct cytobank_experiment object from ACS file
}
