% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowJoWorkspace_Methods.R
\name{open_flowjo_xml}
\alias{open_flowjo_xml}
\alias{closeWorkspace}
\alias{flowjo_ws_close}
\title{Open/Close a flowJo workspace}
\usage{
open_flowjo_xml(file, options = 0, ...)

closeWorkspace(workspace)

flowjo_ws_close(workspace)
}
\arguments{
\item{file}{Full path to the XML flowJo workspace file.}

\item{options}{xml parsing options passed to \code{\link{xmlTreeParse}}}

\item{...}{other arguments passed to \code{\link{xmlTreeParse}}}

\item{workspace}{A \code{flowjo_workspace}}
}
\value{
a \code{flowjo_workspace} object.
}
\description{
Open a flowJo workspace and return a \code{flowjo_workspace} object.
Close a flowjo_workspace, destroying the internal representation of the XML document, and freeing the associated memory.
}
\details{
Open an XML flowJo workspace file and return a \code{flowjo_workspace} object. The workspace is represented using a \code{XMLInternalDocument} object.
	Close a flowJoWorkpsace after finishing with it. This is necessary to explicitly clean up the C-based representation of the XML tree. (See the XML package).
}
\examples{
\dontrun{
	file<-"myworkspace.xml"
	ws<-open_flowjo_xml(file);
	class(ws); #flowjo_workspace
	flowjo_ws_close(ws);
}

}
