% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva2GatingSet.R
\name{openDiva}
\alias{openDiva}
\title{open Diva xml workspace}
\usage{
openDiva(file, options = 0, ...)
}
\arguments{
\item{file}{xml file}

\item{options}{argument passed to \link{xmlTreeParse}}

\item{...}{arguments passed to \link{xmlTreeParse}}
}
\value{
a \code{divaWorkspace} object
}
\description{
open Diva xml workspace
}
\examples{
\dontrun{
library(flowWorkspace)
library(CytoML)
ws <- openDiva(system.file('extdata/diva/PE_2.xml', package = "flowWorkspaceData"))
ws
getSampleGroups(ws)
getSamples(ws)
gs <- parseWorkspace(ws, name = 2, subset = 1)
sampleNames(gs)
getNodes(gs)
plotGate(gs[[1]])
}
}
