% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range.GatingHierarchy}
\alias{range.GatingHierarchy}
\title{the parameter range from the flow data associated with GatingHierarchy}
\usage{
\method{range}{GatingHierarchy}(..., na.rm = FALSE, type = c("instrument", "data"), raw.scale = FALSE)
}
\arguments{
\item{...}{GatingHierarchy object}

\item{na.rm}{not used}

\item{type}{character of "instrument" or "data" indicating whether to retrieve the instrument or the actual data range}

\item{raw.scale}{logical whether convert the range from transformed scale to raw scale}
}
\value{
matrix
}
\description{
the parameter range from the flow data associated with GatingHierarchy
}
\examples{
\dontrun{
 range(gh, type = "data")#return data range
 range(gh) #return instrument range
 range(gh, raw.scale = TRUE) #inverse transform the range to the raw scale
}
}
