% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGML_methods.R
\name{show,graphGML-method}
\alias{show,graphGML-method}
\title{show method for graphGML}
\usage{
\S4method{show}{graphGML}(object)
}
\arguments{
\item{object}{\code{graphGML}}
}
\value{
nothing
}
\description{
show method for graphGML
}
