% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobank2GatingSet.R
\name{compare.counts}
\alias{compare.counts}
\title{compare the counts to cytobank's exported csv so that the parsing result can be verified.}
\usage{
\method{compare}{counts}(gs, file, id.vars = c("FCS Filename", "population"),
  ...)
}
\arguments{
\item{gs}{parsed GatingSet}

\item{file}{the stats file (contains the populatio counts) exported from cytobank.}

\item{id.vars}{either "population" or "FCS filename" that tells whether the stats file format is one population per row or FCS file per row.}

\item{...}{arguments passed to data.table::fread function}
}
\value{
a data.table (in long format) that contains the counts from openCyto and Cytobank side by side.
}
\description{
compare the counts to cytobank's exported csv so that the parsing result can be verified.
}
\examples{

xmlfile <- system.file("extdata/cytotrol_tcell_cytobank.xml", package = "CytoML")
fcsFiles <- list.files(pattern = "CytoTrol", system.file("extdata", package = "flowWorkspaceData"), full = TRUE)
gs <- cytobank2GatingSet(xmlfile, fcsFiles)
## verify the stats are correct
statsfile <- system.file("extdata/cytotrol_tcell_cytobank_counts.csv", package = "CytoML")
dt_merged <- compare.counts(gs, statsfile, id.vars = "population")
all.equal(dt_merged[, count.x], dt_merged[, count.y], tol = 5e-4)

}
