% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobank2GatingSet.R, R/cytobankExperiment.R
\name{cytobank2GatingSet.default}
\alias{cytobank2GatingSet.default}
\alias{cytobank2GatingSet.cytobankExperiment}
\title{A wrapper that parse the gatingML and FCS files (or cytobankExperiment object) into GatingSet}
\usage{
\method{cytobank2GatingSet}{default}(x, FCS)

\method{cytobank2GatingSet}{cytobankExperiment}(x, ...)
}
\arguments{
\item{x}{the cytobankExperiment object or the full path of gatingML file}

\item{FCS}{FCS files to be loaded}
}
\value{
a GatingSet
}
\description{
A wrapper that parse the gatingML and FCS files (or cytobankExperiment object) into GatingSet
}
\examples{
\dontrun{
xmlfile <- system.file("extdata/cytotrol_tcell_cytobank.xml", package = "CytoML")
fcsFiles <- list.files(pattern = "CytoTrol", 
      system.file("extdata", package = "flowWorkspaceData"), full = TRUE)
gs <- cytobank2GatingSet(xmlfile, fcsFiles)
#plotGate(gs[[1]])
}
}
