% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobank2GatingSet.R
\name{cytobank2GatingSet}
\alias{cytobank2GatingSet}
\title{A wrapper that parse the gatingML and FCS files into GatingSet}
\usage{
cytobank2GatingSet(xml, FCS)
}
\arguments{
\item{xml}{the full path of gatingML file}

\item{FCS}{FCS files to be loaded}
}
\value{
a GatingSet
}
\description{
A wrapper that parse the gatingML and FCS files into GatingSet
}
\examples{

xmlfile <- system.file("extdata/cytotrol_tcell_cytobank.xml", package = "CytoML")
fcsFiles <- list.files(pattern = "CytoTrol", 
      system.file("extdata", package = "flowWorkspaceData"), full = TRUE)
gs <- cytobank2GatingSet(xmlfile, fcsFiles)
#plotGate(gs[[1]])

}
