% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\name{cytobank_experiment-methods}
\alias{cytobank_experiment-methods}
\alias{print.cytobank_experiment}
\alias{markernames,cytobank_experiment-method}
\alias{colnames,cytobank_experiment-method}
\alias{sampleNames,cytobank_experiment-method}
\alias{pData,cytobank_experiment-method}
\title{Methods for interacting with cytobank_experiment objects}
\usage{
\S4method{markernames}{cytobank_experiment}(object)

\S4method{colnames}{cytobank_experiment}(x, do.NULL = "missing", prefix = "missing")

\S4method{sampleNames}{cytobank_experiment}(object)

\S4method{pData}{cytobank_experiment}(object)
}
\arguments{
\item{object}{A \code{cytobank_experiment} object}

\item{x}{cytobank_experiment}

\item{do.NULL, prefix}{not used}
}
\description{
These methods mirror similar accessor methods for the \code{GatingSet}
class.
}
