% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva2GatingSet.R
\docType{class}
\name{diva_workspace-class}
\alias{diva_workspace-class}
\alias{diva_get_sample_groups}
\alias{diva_get_samples}
\alias{show,diva_workspace-method}
\alias{parseWorkspace,diva_workspace-method}
\alias{diva_to_gatingset}
\title{diva_workspace class
Inherited from \link{flowjo_workspace-class}}
\usage{
diva_get_sample_groups(x)

diva_get_samples(x)

\S4method{show}{diva_workspace}(object)

\S4method{parseWorkspace}{diva_workspace}(obj, ...)

diva_to_gatingset(obj, name = NULL, subset = NULL, path = obj@path,
  fast = TRUE, worksheet = c("normal", "global"),
  swap_cols = list(`FSC-H` = "FSC-W", `SSC-H` = "SSC-W"),
  verbose = FALSE, ...)
}
\arguments{
\item{x}{diva_workspace}

\item{object}{diva_workspace}

\item{obj}{diva_workspace}

\item{...}{other arguments}
}
\description{
diva_workspace class
Inherited from \link{flowjo_workspace-class}
}
