% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva2GatingSet.R
\docType{class}
\name{diva_workspace-class}
\alias{diva_workspace-class}
\alias{show,diva_workspace-method}
\title{An R representation of a BD FACSDiva workspace}
\description{
Inherited from \link{flowjo_workspace-class}
}
\section{Slots}{
 
\describe{
  \item{\code{version}:}{Object of class \code{"character"}. The version of the XML workspace. }
  \item{\code{file}:}{Object of class \code{"character"}. The file name. }
  \item{\code{.cache}:}{Object of class \code{"environment"}. An environment for internal use.  }
	\item{\code{path}:}{Object of class \code{"character"}. The path to the file. }
  \item{\code{doc}:}{Object of class \code{"XMLInternalDocument"}. The XML document object. }
  \item{\code{options}:}{Object of class \code{"integer"}. The XML parsing options passed to \code{\link{xmlTreeParse}}. }
  }
}

\seealso{
\code{\linkS4class{GatingSet}} 
  \code{\linkS4class{GatingHierarchy}}
}
