% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva2GatingSet.R
\docType{class}
\name{divaWorkspace-class}
\alias{divaWorkspace-class}
\title{divaWorkspace class
Inherited from \link{flowJoWorkspace}}
\description{
divaWorkspace class
Inherited from \link{flowJoWorkspace}
}

