% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGML_methods.R
\name{getChildren,graphGML,character-method}
\alias{getChildren,graphGML,character-method}
\title{get children nodes}
\usage{
\S4method{getChildren}{graphGML,character}(obj, y)
}
\arguments{
\item{obj}{\code{graphGML}}

\item{y}{\code{character} parent node path}
}
\value{
a graphNEL node
}
\description{
get children nodes
}
\examples{
\dontrun{
g <- read.gatingML.cytobank(xmlfile)
getChildren(g, "GateSet_722326")
getParent(g, "GateSet_722326")
}
}
