% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobankExperiment.R
\name{open_cytobank_experiment}
\alias{open_cytobank_experiment}
\alias{cytobankExperiment}
\title{Construct a \code{cytobank_experiment} object from ACS file}
\usage{
open_cytobank_experiment(acs, exdir = tempfile())
}
\arguments{
\item{acs}{ACS file exported from Cytobank}

\item{exdir}{he directory to extract files to}
}
\value{
cytobank_experiment object
}
\description{
Construct a \code{cytobank_experiment} object from ACS file
}
